+++ xptdr/CGameOver.h
@@ -176,8 +176,8 @@ inline void CGameOver<MType>::updateCharacter()
 	for (int i = 1; i < charactersInfo.size(); i++) {
 		if (charactersInfo[i][0] == character.getName()) {
 
-			int maxHealthPointTemp = 20;
-			if (charactersInfo[i][4] != "" && typeid(std::stoi(charactersInfo[i][4])) == typeid(int)) maxHealthPointTemp = std::stoi(charactersInfo[i][4]);
+			float maxHealthPointTemp = 20;
+			if (charactersInfo[i][4] != "" && typeid(std::stof(charactersInfo[i][4])) == typeid(int)) maxHealthPointTemp = std::stof(charactersInfo[i][4]);
 
 			float moveSpeedTemp = 0.5f;
 			if (charactersInfo[i][5] != "" && typeid(std::stof(charactersInfo[i][5])) == typeid(float)) moveSpeedTemp = std::stof(charactersInfo[i][5]);
